Add files via upload
I am not political, I do not feel good about this, this is none of my business. These were released to the public. This is all legal. I do not pay attention to politics and I do not know any of these people on a personal level. If any single person wants me to take this down, contact me. If you really need it to be dropped, then you will find a way to contact me and I will gladly do so.